refactor: update links module
