+++ src/shared/cell-content/source-file-content.ts
@@ -1,5 +1,3 @@
-import { getBasename } from "../link/link-utils";
-
 export const getSourceFileContent = (
 	content: string,
 	shouldRemoveMarkdown = false
@@ -7,7 +5,5 @@ export const getSourceFileContent = (
 	if (shouldRemoveMarkdown) return content;
 	if (content === "") return "";
 
-	const basename = getBasename(content);
-
-	return `[[${content}|${basename}]]`;
+	return `[[${content}]]`;
 };
