+++ src/shared/link-and-path/constants.ts
@@ -1,0 +1,1 @@
+export const MARKDOWN_EXTENSION = "md";
