+++ src/shared/loom-state/commands/column-add-command/utils.ts
@@ -31,7 +31,7 @@ export const columnAddExecute = (
 	});
 
 	return {
-		nextColumns,
-		nextRows,
+		columns: nextColumns,
+		rows: nextRows,
 	};
 };
